rename the entry point `pinboard.main.tsx` to `app.tsx` to avoid confusion with the component `pinboard.tsx`
